requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
scikit-image==0.21.0
Pillow==10.0.0
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
albumentations==1.3.1
PyYAML==6.0.1
tqdm==4.66.1
plotly==5.17.0 